Extract workflow validate diagnostics
